requirements.txt
anthropic==0.29.0
chromadb==0.5.3
ffmpeg==1.4
flask==3.0.3
Flask-Cors==4.0.1
firebase-admin==6.5.0
jsonpickle==3.2.2
google-cloud-aiplatform==1.68.0
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.25.0
google-cloud-discoveryengine==0.11.13
google-cloud-firestore==2.17.0
google-cloud-storage==2.17.0
google-cloud-tasks==2.16.4
google-cloud-texttospeech==2.16.3
langchain==0.2.6
langchain-community==0.2.6
langchainhub==0.1.21
langchain-google-vertexai==1.0.6
nltk==3.8.1
openai==1.35.4
Random-Word==1.0.11
pydub==0.25.1
python-dotenv==1.0.1
python-magic==0.4.27
sqlalchemy==2.0.31
sqlalchemy-bigquery==1.11.0
tabulate==0.9.0
tenacity==8.4.2
tiktoken==0.7.0
